+++ Services/ServiceBase.cs
@@ -1,4 +1,5 @@
-﻿using Assets.Scripts.Managers;
+﻿using Assets.Scripts.Game.Other;
+using Assets.Scripts.Managers;
 
 namespace BonkTuner.Services;
 
@@ -7,5 +8,10 @@ internal abstract class ServiceBase
     public static bool CanContinue()
     {
         return ConfigService.Core.IsEnabled && (ConfigService.Core.IsEnabledDuringChallenges || MapController.runConfig?.challenge == null);
+    }
+
+    public static bool CanContinue(RunConfig newRunConfig)
+    {
+        return ConfigService.Core.IsEnabled && (ConfigService.Core.IsEnabledDuringChallenges || newRunConfig?.challenge == null);
     }
 }